LayoutTests/crypto/subtle/rsaes-pkcs1-v1_5-generate-key-extractable-expected.txt

Test generating an RSA key using RSAES-PKCS1-v1_5 algorithm.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Generating a key pair...
PASS keyPair.toString() is '[object Object]'
PASS keyPair.publicKey.type is 'public'
PASS keyPair.publicKey.extractable is true
PASS keyPair.publicKey.algorithm.name is 'RSAES-PKCS1-v1_5'
PASS keyPair.publicKey.algorithm.modulusLength is 2048
PASS bytesToHexString(keyPair.publicKey.algorithm.publicExponent) is '010001'
PASS keyPair.publicKey.algorithm.hash is undefined.
PASS keyPair.publicKey.usages is ['encrypt']
PASS keyPair.privateKey.type is 'private'
PASS keyPair.privateKey.extractable is true
PASS keyPair.privateKey.algorithm.name is 'RSAES-PKCS1-v1_5'
PASS keyPair.privateKey.algorithm.modulusLength is 2048
PASS bytesToHexString(keyPair.privateKey.algorithm.publicExponent) is '010001'
PASS keyPair.privateKey.algorithm.hash is undefined.
PASS keyPair.privateKey.usages is ['decrypt']
PASS successfullyParsed is true

TEST COMPLETE
